Improve code that adjusts mob spawner characteristics
I was hoping to make the changes non-persistent, such that if the mod was
removed, the spanwers would revert to normal behavior, however the older code
was not the correct way of doing things, and only worked when the world was
reloaded. THis changes the event used, and changes spawner characteristics via
modifying the NBT data of the spawner.