Refactor spawn management code... again
Cancelling the potential spawn event prevented Extra Utilities cursed earth
from spawning mobs. I want to allow that to continue to work because its
something that you as the player choose to create/place, so you should be
ready for dealing with the mobs that will spawn.

The current model unfortunately requires me to add exceptions for particular
circumstances, so for now, vanilla mob spawners and Extra Utilities cursed
earth are the only blocks that will automatically spawn mobs in the world.

I fully intend to add support for MFR auto-spawners and Ender IO powered
spawners in the near future. I'd also like to try and add some kind of
exception list in the configs, but it all depends on how mods spawn mobs.